Improve hashbangs
This is meant to be run in a virtual environment, and the readme
documents that, so python (rather than python3) is the typical, and
slightly more reliable, way to go. Also, these days, python3 is not
necessarily safer than python, which no longer tends to refer to
python2 and which is often the only available command.