Re-run notebooks
This tests the updated dependencies, but the improved timings are
most likely not mostly the result of them. I think it's been a
while since I've re-run these notebooks (and committed the changed
results). I think the platform and Python version are the most
important factors in improving the timing.