Add Python versions to CI
And enable `fail-fast: false` in more places, perhaps temporarily,
to figure out all of which versions/platforms are failing, why they
are failing, and which of them will make sense to try to fix.

I'm doing this now with the goal of allowing CI to pass on
Dependabot PRs, especially security update PRs like #230, while
being able to be pretty confident that they don't break anything.