Remove CodeQL workflow (in prep for default setup)
This removes the CodeQL workflow, because I am going to try
enabling the "default" setup. The default setup works pretty well
these days, avoids having to update actions or make changes to
account for drift in recommended configuration or to remove steps
that become unnecessary over time (such as installing dependencies,
for some languages), and makes it easy to expand scanning to also
cover GitHub Actions workflows (which is now fully supported).