Drop non-locked `pipenv` test on Windows entirely
`pipenv` is picking the 32-bit Python when it should prefer the
architecture-matching 64-bit Python. I don't know why this happens.
It looks like it might be a bug in `pipenv` but I am not sure.

This isn't worth fixing unless development is to be resumed and
alternatives to `pipenv` (such as `uv`) turn out to be unsuitable.
Instead, this commit drops the Windows jobs in `pytest-pipenv`. All
the other Windows jobs are kept: the locked `pipenv` test job for
Windows in `pytest-pipenv-lock`, as well as the seven Windows
`conda` jobs in `pytest-conda`.